Let user choose number of cells; reorganize code
Implemented function to let the user choose the number of cells per
row and column (and hence the total number of cells). This required
first implementing a function to delete all child nodes from the
container div first. Also abstracted out much of the code that was in
the global space into functions in order to effectively redraw the
grid and improve code readability.